% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine.R
\name{stop_runner}
\alias{stop_runner}
\title{Stop execution and reset the runner (Stop button)}
\usage{
stop_runner(runner, run_log)
}
\arguments{
\item{runner}{Runner from \code{make_runner()}.}

\item{run_log}{A \code{reactiveVal} for log output.}
}
\description{
Stop execution and reset the runner (Stop button)
}
